Extract conference carry-over logic into portal/services (no inline imports)
The carry-over/snapshot operations were methods on Conference, which forced
function-level imports of the sponsorship/volunteer apps to avoid a circular
import (those apps import portal.models). Move them to portal/services.py — a
module nothing imports during model loading — so every cross-app import is at
the top of the file.

- portal/services.py: clone_teams, clone_sponsorship_tiers,
  bring_forward_volunteers, freeze_stats as plain functions (target, source).
- Conference loses those four methods; admin actions and the start-new-year
  wizard call the service functions instead.
- Tests moved to tests/portal/test_services.py with top-level imports; removed
  the inline imports from the portal view/admin tests too.

442 tests pass, 100% coverage; lint clean; no schema change.

Refs #321 #341
