+++ portal/models.py
@@ -94,95 +94,3 @@ class Conference(BaseModel):
     def get_active(cls):
         """Return the active conference, or ``None`` if none is set."""
         return cls.objects.filter(is_active=True).first()
-
-    def clone_teams_from(self, source):
-        """Copy the team structure from another edition into this one.
-
-        Copies each team's name, description, and open-to-new-members flag but
-        not its team leads (those are per-edition volunteers). Teams whose name
-        already exists in this edition are skipped, so it is safe to run more
-        than once. Returns the number of teams created.
-        """
-        from volunteer.models import Team
-
-        existing = set(self.teams.values_list("short_name", flat=True))
-        created = 0
-        for team in source.teams.all():
-            if team.short_name in existing:
-                continue
-            Team.objects.create(
-                conference=self,
-                short_name=team.short_name,
-                description=team.description,
-                open_to_new_members=team.open_to_new_members,
-            )
-            created += 1
-        return created
-
-    def freeze_stats(self):
-        """Snapshot this edition's live metrics into ``historical_snapshot``.
-
-        Once frozen, the comparison and historical-fallback views read these
-        fixed numbers instead of recomputing them, so a closed edition's stats
-        stay final. Amounts are stored as floats so the dict is JSON-safe.
-        Returns the snapshot dict.
-        """
-        from portal.common import (
-            get_attendee_count_cache,
-            get_donors_count_cache,
-            get_sponsorship_committed_amount_stats_cache,
-            get_sponsorship_committed_count_stats_cache,
-            get_total_donations_amount_cache,
-        )
-
-        self.historical_snapshot = {
-            "registrations": get_attendee_count_cache(self),
-            "sponsors": get_sponsorship_committed_count_stats_cache(self),
-            "sponsorship_amount": float(
-                get_sponsorship_committed_amount_stats_cache(self)
-            ),
-            "donors": get_donors_count_cache(self),
-            "donation_amount": float(get_total_donations_amount_cache(self)),
-        }
-        self.save()
-        return self.historical_snapshot
-
-    def clone_sponsorship_tiers_from(self, source):
-        """Copy sponsorship tiers (name, amount, description) from another edition.
-
-        Skips tiers whose name already exists in this edition. Returns the
-        number of tiers created.
-        """
-        from sponsorship.models import SponsorshipTier
-
-        existing = set(self.sponsorship_tiers.values_list("name", flat=True))
-        created = 0
-        for tier in source.sponsorship_tiers.all():
-            if tier.name in existing:
-                continue
-            SponsorshipTier.objects.create(
-                conference=self,
-                name=tier.name,
-                amount=tier.amount,
-                description=tier.description,
-            )
-            created += 1
-        return created
-
-    def bring_forward_volunteers_from(self, source):
-        """Bring every APPROVED volunteer from another edition into this one.
-
-        Each becomes a PENDING profile (see
-        ``VolunteerProfile.bring_forward_to``); volunteers who already have a
-        profile here are skipped. Returns the number of profiles created.
-        """
-        from volunteer.constants import ApplicationStatus
-
-        created = 0
-        approved = source.volunteer_profiles.filter(
-            application_status=ApplicationStatus.APPROVED
-        )
-        for profile in approved:
-            if profile.bring_forward_to(self) is not None:
-                created += 1
-        return created
